requirements.txt
Flask==2.3.2
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
urllib3<2.0
ics>=0.7.2
psycopg2-binary==2.9.10

